pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "assessment-generator"
version = "1.0.0"
description = "AI-powered MCQ assessment generator with FastAPI backend"
authors = [{name = "Satyam Raj"}]
license = {text = "MIT"}
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "openai>=1.0.0",
    "python-docx>=1.1.0",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "httpx>=0.27.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
]

[tool.setuptools.packages.find]
include = ["assessment_engine*"]
